refactor: 인스턴스, 메서드 분리
